requirements_test.txt
anyconfig
babel
Flask
Flask-Cache
Flask-Analytics
keen
loggly-handler
newrelic
redis
requests
sendgrid
twilio

Jinja2
MarkupSafe
Werkzeug
itsdangerous
wsgiref

# test
pytest
sure
vcrpy